DOM Overlay: update API to match current draft spec
The current draft spec https://immersive-web.github.io/dom-overlays/ has
several API shape changes compared to the currently-implemented
prototype, and this CL updates the implementation to match the new API.

Instead of asking applications to use the Fullscreen API to change
the active DOM overlay element, use a separate initialization dictionary
to configure this at session creation time, and make that unmodifiable
from that point on until the session ends.

Also add a domOverlayState XRSession attribute that shows that the
mode is working (this isn't obvious if it was requested as an optional
feature), and also the type of overlay (screen or floating).

This change is part of a series. The following parts are logically also
part of the API update, but split out to keep the size more manageable.
The next steps include renaming the feature name and implementing the
new event deduplication logic.

Bug: 991747
Change-Id: I032fba53f0180d791b0999ed93393bfb4467b3cd