Reject cookies with empty names and values.
As discussed in [1], cookies with empty names and empty values should be
rejected. This patch removes the carveout made in https://crbug.com/601786,
and adjusts unittests accordingly.

This patch does not change the WPT expectations; we'll do that at the same
time we change the spec. In the meantime, we'll check in local expectations
matching the behavior we believe is correct.

[1]: https://github.com/httpwg/http-extensions/issues/159#issuecomment-569233866

Bug: 1037996, 601786
Change-Id: I53319cee385efff019b313479184236c53b1d783